requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt
python-dotenv